fastapi/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
fastembed>=0.4.0
faiss-cpu>=1.12.0
numpy>=1.26.0
httpx>=0.27.0
python-dotenv>=1.0.0
slowapi>=0.1.9
